requirements.txt
streamlit
pandas
geopandas
plotly
folium
streamlit_folium
shapely.geometry
geopandas.tools

